NOTICKET // remove base dockerfile and add interactions for button
