+++ docker/DOCKERFILE.storybook
@@ -1,4 +1,11 @@
-FROM neo-materia-base:latest
+FROM node:20-alpine
+
+WORKDIR /app
+
+COPY package.json package-lock.json ./
+RUN npm ci
+
+COPY . .
 
 EXPOSE 3000
 CMD ["npm", "run", "storybook"]
