+++ docker/DOCKERFILE.test
@@ -1,3 +1,10 @@
-FROM neo-materia-base:latest
+FROM node:20-alpine
+
+WORKDIR /app
+
+COPY package.json package-lock.json ./
+RUN npm ci
+
+COPY . .
 
 CMD ["npm", "run", "test:unit"]
