+++ src/components/01-atoms/Button/Button.stories.ts
@@ -1,4 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/vue3'
+
+import { expect, fn, userEvent, waitFor, within } from '@storybook/test'
+
 import Button from './Button.vue'
 
 const meta: Meta<typeof Button> = {
@@ -13,6 +16,7 @@ const meta: Meta<typeof Button> = {
     label: 'Click me',
     size: 'medium',
     disabled: false,
+    onClick: fn(),
   },
 }
 
@@ -30,5 +34,18 @@ export const Disabled: Story = {
   args: {
     label: 'Disabled Button',
     disabled: true,
+  },
+}
+
+export const interaction: Story = {
+  play: async ({ args, canvasElement, step }) => {
+    const canvas = within(canvasElement)
+
+    await step('Check click functionality', async () => {
+      await userEvent.click(canvas.getByRole('button'))
+    })
+
+    // 👇 Now we can assert that the onSubmit arg was called
+    await waitFor(() => expect(args.onClick).toHaveBeenCalled())
   },
 }
